main.js разделен на модули Внесены исправления по рекомендациям к прошлому Pull Resuest Внесены исправления в функции получения случайных значений
